+++ components/ui/icon/index.tsx
@@ -91,7 +91,8 @@ const iconStyle = tva({
       sm: 'h-4 w-4',
       md: 'h-[18px] w-[18px]',
       lg: 'h-5 w-5',
-      xl: 'h-6 w-6'
+      xl: 'h-6 w-6',
+      '2xl': 'h-7 w-7'
     }
   }
 })