+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Changed
 
 - Replace temporary menu components in user and settings screen ♻️
+- Adjust button and avatar sizes 📏
 
 ### Fixed